Right-Left Cipher
An encoded string is given which have to be decoded. Place the middle character as the first character of the resulted string. Then print first right character of the middle character and then the first left character. Do it as a pair until right side became equal or equal to given string length and left side greater or equal to zero.